Master Anim Layer Part 3 Done
Kinda fixed the issue of the Blend Space Legacy not going in the Anim Layer but not really but it works lol. How i fixed it was I retarged the legacy bp to the cone, basically retargeting it to itself and it fixed it kinda but when i deleted the unused files, it went back to not adding it in there but I redo it and it stays there so I wont switch it lol.